Adding example send message calls and a text prompt to stop the bot
Figured out how to send messages to a specified channel, stuck those
calls in for reference.

Got the channels from a server for reference.

Added a prompt to stop the bot. Not sure if we can just wait on this
input indefinitely and still have the handlers work like normal, will
require testing out when we start putting in handlers.

Updated the README with some features I'm thinking of.