Added a couple event handlers
This is a messy commit, basically just saving my progress and sharing
what I've learned so far.

Originally I thought that the Recipients list on the Channel object
would contain all the users in that channel, but that doesn't seem to be
the case.

Each server (guild) has a list of VoiceState objects, one for each
member of the guild (at least the ones in a voice channel). The VoiceState
object contains which voice channel the corresponding member is in. This
is the object to watch, through the VoiceStateUpdate event.

Once the functionality that we want is there (or maybe sooner if I feel
like it), we'll refactor to make the code not stink as bad (huge
functions, all in the same file, lots of debug code, etc).